refactor(consensus): replace ad-hoc committer test files with mysticeti suite
`base_committer_tests.rs` and `pipelined_committer_tests.rs` predate the
per-protocol breakdown: the first runs an entirely fictitious
non-pipelined-Mysticeti configuration that doesn't match any production
protocol, and the second covers real `Protocol::mysticeti` with
leader_count=1 but in the old ad-hoc style.

Adds `mysticeti_tests.rs` using the shared per-protocol scaffolding —
closes the gap where the namesake protocol was the only one without a
file in the unified style. Drops both ad-hoc files; the scenarios they
covered are now exercised through the eight per-protocol fixtures
(BlueBottle, Nemo-Nemo, Orcaella, Cordial PS+Async, Mysticeti, Mahi-Mahi
wl=4+wl=5), and the BFT-symmetric wl=3 skip path remains covered by
Mahi-Mahi (`direct_skip_quorum = 2n/3+1`).

`multi_committer_tests.rs` is kept as-is — it's the only coverage for
`leader_count > 1`, which the single-leader shared runners don't
currently support.
